DDF-5793 address comments and move dms validation into validation file
