+++ ui/packages/catalog-ui-search/src/main/webapp/component/location-new/utils/dms-utils.js
@@ -404,106 +404,6 @@ function getSecondsPrecision(dmsCoordinate) {
   }
 }
 
-const lat = {
-  degreesBegin: 0,
-  degreesEnd: 2,
-  minutesBegin: 3,
-  minutesEnd: 5,
-  secondsBegin: 6,
-  secondsEnd: -1,
-}
-const lon = {
-  degreesBegin: 0,
-  degreesEnd: 3,
-  minutesBegin: 4,
-  minutesEnd: 6,
-  secondsBegin: 7,
-  secondsEnd: -1,
-}
-
-const validateDmsLatInput = input => {
-  const degrees = input.slice(lat.degreesBegin, lat.degreesEnd)
-  const minutes = input.slice(lat.minutesBegin, lat.minutesEnd)
-  const seconds = input.slice(lat.secondsBegin, lat.secondsEnd)
-  const maxDmsLat = '90°00\'00"'
-  if (degrees > 90) {
-    return maxDmsLat
-  } else if (minutes >= 60) {
-    if (degrees < 90) {
-      return (Number.parseInt(degrees) + 1).toString() + '°00\'00"'
-    } else {
-      return maxDmsLat
-    }
-  } else if (seconds >= 60) {
-    if (minutes < 59) {
-      return degrees + '°' + (Number.parseInt(minutes) + 1).toString() + '\'00"'
-    } else {
-      if (degrees >= '90') {
-        return maxDmsLat
-      } else {
-        return (Number.parseInt(degrees) + 1).toString() + '°00\'00"'
-      }
-    }
-  } else if (
-    input.slice(lat.degreesBegin, lat.degreesEnd) === '9_' &&
-    input.slice(lat.degreesEnd) === '°00\'00"'
-  ) {
-    return '9_°__\'__"'
-  } else if (
-    input.slice(lat.minutesBegin, lat.minutesEnd) === '6_' &&
-    input.slice(lat.minutesEnd) === '\'00"'
-  ) {
-    return input.slice(lat.degreesBegin, lat.degreesEnd) + '°6_\'__"'
-  } else {
-    return input
-  }
-}
-
-const validateDmsLonInput = input => {
-  const degrees = input.slice(lon.degreesBegin, lon.degreesEnd)
-  const minutes = input.slice(lon.minutesBegin, lon.minutesEnd)
-  const seconds = input.slice(lon.secondsBegin, lon.secondsEnd)
-  const maxDmsLon = '180°00\'00"'
-  if (degrees > 180) {
-    return maxDmsLon
-  } else if (minutes >= 60) {
-    if (degrees < 180) {
-      return (Number.parseInt(degrees) + 1).toString() + '°00\'00"'
-    } else {
-      return maxDmsLon
-    }
-  } else if (seconds > 60) {
-    if (minutes < 59) {
-      return degrees + '°' + (Number.parseInt(minutes) + 1).toString() + '\'00"'
-    } else {
-      if (degrees >= '180') {
-        return maxDmsLon
-      } else {
-        return (Number.parseInt(degrees) + 1).toString() + '°00\'00"'
-      }
-    }
-  } else if (
-    input.slice(lon.degreesBegin, lon.degreesEnd) === '18_' &&
-    input.slice(lon.degreesEnd) === '°00\'00"'
-  ) {
-    return '18_°__\'__"'
-  } else if (
-    input.slice(lon.minutesBegin, lon.minutesEnd) === '6_' &&
-    input.slice(lon.minutesEnd) === '\'00"'
-  ) {
-    return input.slice(lon.degreesBegin, lon.degreesEnd) + '°6_\'__"'
-  } else {
-    return input
-  }
-}
-const validateInput = (input, placeHolder) => {
-  if (input !== undefined && placeHolder === 'dd°mm\'ss.s"') {
-    return validateDmsLatInput(input)
-  } else if (input !== undefined && placeHolder === 'ddd°mm\'ss.s"') {
-    return validateDmsLonInput(input)
-  }
-}
-
 module.exports = {
   dmsToWkt,
   validateDms,
@@ -514,5 +414,4 @@ module.exports = {
   ddToDmsCoordinateLon,
   getSecondsPrecision,
   Direction,
-  validateInput,
 }
