+++ ui/packages/catalog-ui-search/src/main/webapp/react-component/location/location.js
@@ -55,7 +55,7 @@ const inputs = plugin({
           setState={({ value, ...data }) => {
             setState({ keywordValue: value, ...data })
           }}
-          setBufferState={(key, value) => setState(key, value)}
+          setBufferState={(key, value) => setState({ [key]: value })}
         />
       )
     },
@@ -88,7 +88,7 @@ const LocationInput = props => {
     <Root isOpen={input.label !== undefined}>
       <Component>
         <Dropdown label={input.label || 'Select Location Option'}>
-          <Menu value={mode} onChange={value => setState('mode', value)}>
+          <Menu value={mode} onChange={value => setState({['mode']: value})}>
             {Object.keys(inputs).map(key => (
               <MenuItem key={key} value={key}>
                 {inputs[key].label}
