+++ ui/packages/catalog-ui-search/src/main/webapp/react-component/location/point-radius.js
@@ -211,6 +211,7 @@ const PointRadiusUsngMgrs = props => {
       >
         <TextField
           label="Radius"
+          type='number'
           value={String(radius)}
           onChange={value => {
             setRadiusError(validateGeo('radius', value))
@@ -329,6 +330,7 @@ const PointRadiusUtmUps = props => {
       >
         <TextField
           label="Radius"
+          type='number'
           value={String(radius)}
           onChange={value => {
             setRadiusError(validateGeo('radius', value))
